changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11732 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Cocoa Extensions/NSArray+Extensions.h
@@ -44,8 +44,8 @@
 @interface NSMutableArray (OrcaExtensions)
 - (void) insertObjectsFromArray:(NSArray *)array atIndex:(int)index;
 - (NSMutableArray*) children;
-- (unsigned) numberOfChildren;
-- (void) moveObject:(id)anObj toIndex:(unsigned)newIndex;
+- (NSUInteger) numberOfChildren;
+- (void) moveObject:(id)anObj toIndex:(NSUInteger)newIndex;
 - (void)shuffle;
 
 //implements stack behavior
